+++ src/Symfony/Component/Form/Extension/Validator/ValidatorTypeGuesser.php
@@ -122,7 +122,13 @@ class ValidatorTypeGuesser implements FormTypeGuesserInterface
 
             case 'Symfony\Component\Validator\Constraints\File':
             case 'Symfony\Component\Validator\Constraints\Image':
-                return new TypeGuess('Symfony\Component\Form\Extension\Core\Type\FileType', [], Guess::HIGH_CONFIDENCE);
+                return new TypeGuess(
+                    'Symfony\Component\Form\Extension\Core\Type\FileType',
+                    ($constraint->mimeTypes && \is_array($constraint->mimeTypes)) ?
+                        ['attr' => ['accept' => implode(', ', $constraint->mimeTypes)]] :
+                        [],
+                    Guess::HIGH_CONFIDENCE
+                );
 
             case 'Symfony\Component\Validator\Constraints\Language':
                 return new TypeGuess('Symfony\Component\Form\Extension\Core\Type\LanguageType', [], Guess::HIGH_CONFIDENCE);