+++ src/Symfony/Component/Form/Tests/Extension/Validator/ValidatorTypeGuesserTest.php
@@ -16,6 +16,7 @@ use Symfony\Component\Form\Extension\Validator\ValidatorTypeGuesser;
 use Symfony\Component\Form\Guess\Guess;
 use Symfony\Component\Form\Guess\ValueGuess;
 use Symfony\Component\Validator\Constraints\Email;
+use Symfony\Component\Validator\Constraints\File;
 use Symfony\Component\Validator\Constraints\IsTrue;
 use Symfony\Component\Validator\Constraints\Length;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -109,6 +110,38 @@ class ValidatorTypeGuesserTest extends TestCase
 
         $result = $this->guesser->guessMaxLengthForConstraint($constraint);
         $this->assertNull($result);
+    }
+
+    /**
+     * Check if FileType contains accept attribute.
+     */
+    public function testGuessMimeTypesForConstraintWithMimeTypesValue()
+    {
+        $mineTypes = ['image/png', 'image/jpeg'];
+        $constraint = new File(['mimeTypes' => $mineTypes]);
+
+        $typeGuess = $this->guesser->guessTypeForConstraint($constraint);
+
+        $this->assertInstanceOf('Symfony\Component\Form\Guess\TypeGuess', $typeGuess);
+        $this->assertArrayHasKey('attr', $typeGuess->getOptions());
+        $this->assertArrayHasKey('accept', $typeGuess->getOptions()['attr']);
+        $this->assertEquals(implode(',', $mineTypes), $typeGuess->getOptions()['attr']['accept']);
+    }
+
+    /**
+     * Check if file assert not contains mime types.
+     */
+    public function testGuessMimeTypesForConstraintWithoutMimeTypesValue()
+    {
+        $constraint = new File();
+
+        $typeGuess = $this->guesser->guessTypeForConstraint($constraint);
+
+        $this->assertInstanceOf('Symfony\Component\Form\Guess\TypeGuess', $typeGuess);
+
+        if (isset($typeGuess->getOptions()['attr'])) {
+            $this->assertArrayNotHasKey('accept', $typeGuess->getOptions()['attr']);
+        }
     }
 
     public function maxLengthTypeProvider()